+++ .gitignore
@@ -1,6 +1,8 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
+.gitignore
+package-lock.json
 /node_modules
 /.pnp
 .pnp.js
