--- photos.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-
-class Photo extends Component {
-    render() {
-        return (
-            <img src="https://images.mid-day.com/images/images/2023/feb/khan-srksas_d.jpg" alt="King Khan">
-            </img>
-        )
-    }
-}
-
-export default Photo;
